ICU-6125 Eliminate a few kilobytes of unused bits in constant data structures.
X-SVN-Rev: 23182

+++ icu4c/source/i18n/csrmbcs.cpp
@@ -1,6 +1,6 @@
 /*
  **********************************************************************
- *   Copyright (C) 2005-2006, International Business Machines
+ *   Copyright (C) 2005-2008, International Business Machines
  *   Corporation and others.  All Rights Reserved.
  **********************************************************************
  */
@@ -19,7 +19,7 @@ U_NAMESPACE_BEGIN
 
 #define min(x,y) (((x)<(y))?(x):(y))
 
-const int32_t commonChars_sjis [] = {
+const uint16_t commonChars_sjis [] = {
 // TODO:  This set of data comes from the character frequency-
 //        of-occurence analysis tool.  The data needs to be moved
 //        into a resource and loaded from there.
@@ -30,7 +30,7 @@ const int32_t commonChars_sjis [] = {
 0x8343, 0x834e, 0x834f, 0x8358, 0x835e, 0x8362, 0x8367, 0x8375, 0x8376, 0x8389,
 0x838a, 0x838b, 0x838d, 0x8393, 0x8e96, 0x93fa, 0x95aa};
 
-const int32_t commonChars_euc_jp[] = {
+const uint16_t commonChars_euc_jp[] = {
 // TODO:  This set of data comes from the character frequency-
 //        of-occurence analysis tool.  The data needs to be moved
 //        into a resource and loaded from there.
@@ -45,7 +45,7 @@ const int32_t commonChars_euc_jp[] = {
 0xa5e5, 0xa5e9, 0xa5ea, 0xa5eb, 0xa5ec, 0xa5ed, 0xa5f3, 0xb8a9, 0xb9d4, 0xbaee,
 0xbbc8, 0xbef0, 0xbfb7, 0xc4ea, 0xc6fc, 0xc7bd, 0xcab8, 0xcaf3, 0xcbdc, 0xcdd1};
 
-const int32_t commonChars_euc_kr[] = {
+const uint16_t commonChars_euc_kr[] = {
 // TODO:  This set of data comes from the character frequency-
 //        of-occurence analysis tool.  The data needs to be moved
 //        into a resource and loaded from there.
@@ -60,7 +60,7 @@ const int32_t commonChars_euc_kr[] = {
 0xc0da, 0xc0e5, 0xc0fb, 0xc0fc, 0xc1a4, 0xc1a6, 0xc1b6, 0xc1d6, 0xc1df, 0xc1f6,
 0xc1f8, 0xc4a1, 0xc5cd, 0xc6ae, 0xc7cf, 0xc7d1, 0xc7d2, 0xc7d8, 0xc7e5, 0xc8ad};
 
-const int32_t commonChars_big5[] = {
+const uint16_t commonChars_big5[] = {
 // TODO:  This set of data comes from the character frequency-
 //        of-occurence analysis tool.  The data needs to be moved
 //        into a resource and loaded from there.
@@ -75,7 +75,7 @@ const int32_t commonChars_big5[] = {
 0xb5a5, 0xb5bd, 0xb5d0, 0xb5d8, 0xb671, 0xb7ed, 0xb867, 0xb944, 0xbad8, 0xbb44,
 0xbba1, 0xbdd1, 0xc2c4, 0xc3b9, 0xc440, 0xc45f};
 
-const int32_t commonChars_gb_18030[] = {
+const uint16_t commonChars_gb_18030[] = {
 // TODO:  This set of data comes from the character frequency-
 //        of-occurence analysis tool.  The data needs to be moved
 //        into a resource and loaded from there.
@@ -90,7 +90,7 @@ const int32_t commonChars_gb_18030[] = {
 0xcfb5, 0xcfc2, 0xcfd6, 0xd0c2, 0xd0c5, 0xd0d0, 0xd0d4, 0xd1a7, 0xd2aa, 0xd2b2,
 0xd2b5, 0xd2bb, 0xd2d4, 0xd3c3, 0xd3d0, 0xd3fd, 0xd4c2, 0xd4da, 0xd5e2, 0xd6d0};
 
-static int32_t binarySearch(const int32_t *array, int32_t len, int32_t value)
+static int32_t binarySearch(const uint16_t *array, int32_t len, uint16_t value)
 {
     int32_t start = 0, end = len-1;
     int32_t mid = (start+end)/2;
@@ -142,7 +142,7 @@ CharsetRecog_mbcs::~CharsetRecog_mbcs()
     // nothing to do.
 }
 
-int32_t CharsetRecog_mbcs::match_mbcs(InputText *det, const int32_t commonChars[], int32_t commonCharsLen) {
+int32_t CharsetRecog_mbcs::match_mbcs(InputText *det, const uint16_t commonChars[], int32_t commonCharsLen) {
     int   singleByteCharCount = 0;
     int   doubleByteCharCount = 0;
     int   commonCharCount     = 0;
